pyproject.toml
[project]
name = "task-updater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "todoist-api-python>=3.1.0",
]
